requirements.txt
crewai
openai
python-dotenv
requests
